requirements.txt
torch 
torchvision
matplotlib
numpy
fastapi
uvicorn[standard]
python-multipart
Pillow
pycocotools

fastapi
uvicorn[standard]
python-multipart
torch
torchvision
Pillow
pycocotools
numpy
opencv-python-headless